+++ console-backend-data-manager/routes/packages.js
@@ -25,7 +25,7 @@
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *-------------------------------------------------------------------------------*/
 
-module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
+module.exports = function(express, logger, config, Q, HTTP, isAuthenticated) {
 
   var router = express.Router();
 
@@ -87,7 +87,7 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
   }
 
   /* Get a list of packages available and deployed. */
-  router.get('/', cors(corsOptions), function(req, res) {
+  router.get('/', isAuthenticated, function(req, res) {
     var promise = Q.all([getAvailablePackages(), getDeployedPackages()]);
     promise.then(function(results) {
       var packages = [];
@@ -167,7 +167,7 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
   });
 
   /* Get a list of deployed packages. */
-  router.get('/deployed', cors(corsOptions), function(req, res) {
+  router.get('/deployed', isAuthenticated, function(req, res) {
     var promise = Q.all([getDeployedPackages()]);
     promise.then(function(results) {
       var packages = [];
@@ -183,7 +183,7 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
   });
 
   /* GET Package status by id. */
-  router.get('/:id/status', cors(corsOptions), function(req, res) {
+  router.get('/:id/status', isAuthenticated, function(req, res) {
     var id = req.params.id;
     var promise = Q.all([getPackageStatus(id)]);
     promise.then(function(results) {
@@ -200,7 +200,7 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
   });
 
   /* Get package information by id. */
-  router.get('/:id', cors(corsOptions), function(req, res) {
+  router.get('/:id', isAuthenticated, function(req, res) {
     var getPackagesDetails = function(id) {
       var deferred = Q.defer();
       var url = config.deployment_manager.host + config.deployment_manager.API.packages + '/' + id;
@@ -237,17 +237,12 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
 
   /**
    * Deploy a package
-   *
-   * @param id  package ID
-   * @return 202 accepted
-   * @return 404 package not deployed in repository
-   * @return 409 package already deployed
-   * @return 500 server error
    */
-  router.options('/:id', cors(corsOptions)); // enable pre-flight request for PUT request
-  router.put('/:id', cors(corsOptions), function(req, res) {
+  router.put('/:id', isAuthenticated, function(req, res) {
     var packageId = req.params.id;
-    var deployAPI = config.deployment_manager.host + config.deployment_manager.API.packages + '/' + packageId;
+    var userName = req.query['user.name'];
+    var deployAPI = config.deployment_manager.host + config.deployment_manager.API.packages + '/' + packageId +
+      '?user.name=' + userName;
     logger.debug("packageId", packageId);
 
     if (packageId && packageId !== "") {
@@ -267,7 +262,7 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
              logger.error(request.method, request.url, "error: ", error.status);
              statusRet = error.status;
            })
-           .then(function(data) { res.status(statusRet).send(data); }, function(data) { res.sendStatus(500);} );
+           .then(function(data) { res.status(statusRet).send(data); }, function() { res.sendStatus(500);});
     } else {
       logger.error("Missing required package id to deploy the package");
       res.sendStatus(500);
@@ -276,16 +271,12 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
 
   /**
    * Undeploy a package
-   *
-   * @param id  package ID
-   * @return 202 accepted
-   * @return 404 package not deployed
-   * @return 500 server error
    */
-  router.options('/:id', cors(corsOptions)); // enable pre-flight request for DELETE request
-  router.delete('/:id', cors(corsOptions), function(req, res) {
+  router.delete('/:id', isAuthenticated, function(req, res) {
     var packageId = req.params.id;
-    var deployAPI = config.deployment_manager.host + config.deployment_manager.API.packages + '/' + packageId;
+    var userName = req.query['user.name'];
+    var deployAPI = config.deployment_manager.host + config.deployment_manager.API.packages + '/' + packageId +
+      '?user.name=' + userName;
 
     if (packageId && packageId !== "") {
       logger.info("Undeploying package " + packageId + " to " + req.body.action);
@@ -304,7 +295,7 @@ module.exports = function(express, logger, cors, corsOptions, config, Q, HTTP){
              logger.error(request.method, request.url, "error: ", error.status);
              statusRet = error.status;
            })
-           .then(function(data) { res.status(statusRet).send(data); }, function(data) { res.sendStatus(500);} );
+           .then(function(data) { res.status(statusRet).send(data); }, function() { res.sendStatus(500);});
     } else {
       logger.error("Missing required package id to undeploy the package");
       res.sendStatus(500);