+++ console-backend-utils/dbManager.js
@@ -27,7 +27,8 @@
 var redis = require('redis');
 var async = require('async');
 
-var redisClient = redis.createClient();
+var redisUrl = process.env.REDIS || 'redis://localhost';
+var redisClient = redis.createClient(redisUrl);
 
 function stripRedisDataIdentifiers(responseArray) {
     // Strip-out the prefix data identifier (e.g. 'metric:' etc.) as client doesn't need to care
@@ -70,7 +71,7 @@ module.exports = function(logger) {
        }
 
       if ((id) && (data)) {
-        if (notification && id.indexOf('metric:') >= 0 && id.indexOf('.health') <= 0) {
+        if (notification && ( id.indexOf('metric:') >= 0 || id.indexOf('topic:') >= 0) && id.indexOf('.health') <= 0) {
           redisClient.hmget(id, ['value'], function(err, reply) {
                 if (reply[0] === data.value) {
                   notification = false;